Prefix natives with "sampgdk_native_" to keep original names
Some natives like SetTimer or GetTickCount must be renamed in order
to link sampgdk successfully on Windows due to those being defined in
various system DLLs. So they were renamed in earlier versions, for
example CreateMenu was MenuCreate, and so on.

Now there are multiple extra headers ending on "-defs.h" that contain
similar macros of the form:

#undef SetTimer
#define SetTimer sampgdk_native_SetTimer

for all native functions. This ultimately can break code that uses
WinAPI functions with the same name, so additional measures must be
taken (e.g. not #include both sampgdk and Windows headers in the
same file).