Fix CMake error during configuration due to PLY debug message
PLY always printed 'Generating LALR tables' to stderr when
cidl.Parser's parse() method was called for the first time.
And CMake thought that was an error output from exec_process()
so it stopped immediately.

The fix is to provide a custom logger to ply.yacc.yacc() that
overloads the debug() method of PlyLogger to suppress all debug
messages generated by yacc.