Fix amx_Allot() returning success on insufficient heap
There is a bug in amx_Allot() where it returns success even though
there's not enough space on the heap:

if (amx->stk - amx->hea - cells*sizeof(cell) < STKMARGIN)
   return AMX_ERR_MEMORY;

The expression on the right side of the comparison is converted
to an unsigned type (because result of sizeof is of type size_t).
and in fact never results in a negative value.