Fix long argument issue of sampgdk_do_vlogprintf()
sampgdk_do_logprintf() didn't work for arguments of size greater than 4
bytes. Now it prints stuff to a temporary buffer instead of passing
arguments directly to logprintf(). As a consequence, you can't have %%
in the format string, but that's not critical (at least not as important
as not being able to print floats!).