Introduce API to intercept public function calls (issue #55)
Make it possible to set a "public hook" that is called whenever some public
AMX function gets called against the gamemode (FSs are ignored).

The hook function takes two parameters - a pointer to the gamemode's AMX
struct, and the name of the public function that is being called. It should
return true if the function is allowed to execute and false if not.