Another attempt to optimize sampgdk_hook_install()
Since we only need to compute the jump offset once it's seems like a
good idea to move the code that calculates it to sampgdk_hook_install()
instead and store the "jump buffer" in the _sampgdk_hook struct.

Also this jump buffer is now 8 bytes instead of 5, so that the compiler
can emit two 32-bit mov instructions instead of a 32-bit mov followed
by an 8-bit one. This doesn't seem to change anything though...