Call amx_Exec() regardless of callback's existence
If the callback is not found in the gamemode, i.e. index == AMX_EXEC_GDK,
amx_Exec() must be called anyway because there may be another plugin running
that needs this. Such is possible if we have two or more plugins one of
which uses the static build or amalgamation.

This bug was apparently introduced in commit
8745448bc5bc9289da30df574e8395a68b0408a6.

Fixes #105.