Add ability to change native search strategy
It is now possible to tell the library to always search for native
functions, even if they previously were not found. This may be useful
in some cases, e.g. debugging broken plugins that try to call natives
too early.

In order to enable this, set the environment variable
SAMPGDK_NATIVE_SEARCH_ALWAYS to 1, e.g.

    SAMPGDK_NATIVE_SEARCH_ALWAYS=1 samp-server-cli -o -d helloworld