Improve handling of special callbacks
Simplify how special callbacks are handled. Now they are not treated as a
special case (well, almost) and get registered like normal callbacks. The
only difference is that their name begins with a colon.

This should fix a bug where _sampgdk_callback_find() would crash inside
bsearch() when registering the fist of two special calblacks  because of
an out-of-bounds array access.

Fixes #169.