Suppress buffer overrun runtime error on Windows in logging functions
sprintf_s and vsprintf_s both issue an error at runtime of they detect
that the destination buffer is too small. We don't want that, instead we
should truncate the resulting string quietly (as is done by
snprintf/vsnprintf on Linux). So we should use the corresponding Windows
API functions.